Message passing and DHT upkeep.
