+++ src/p2p/node.py
@@ -3,7 +3,6 @@ from src.cryptography.rsa import *
 
 from typing import List
 import threading
-import base64
 import socket
 import time
 
@@ -28,9 +27,8 @@ class Node(threading.Thread):
         self.rsa_pub_key = get_rsa_pub_key(True)
 
         # Node & Connection parameters
-        # self.inbound = []
-        # self.outbound = []
-        self.connections = []
+        self.inbound = []
+        self.outbound = []
         self.reconnect = []
 
         self.terminate_flag = threading.Event()
@@ -40,9 +38,9 @@ class Node(threading.Thread):
         # To add ssl encryption?
         # self.sock = ssl.wrap_socket(self.sock)
 
-    # @property
-    # def connections(self):
-    #     return self.inbound + self.outbound
+    @property
+    def connections(self):
+        return self.inbound + self.outbound
 
     def debug_print(self, message):
         if self.debug:
@@ -54,8 +52,8 @@ class Node(threading.Thread):
         self.sock.settimeout(5.0)
         self.sock.listen(1)
 
-    def create_connection(self, connection: socket.socket, id: str, host: str, port: int, parent_port: int = None) -> Connection:
-        return Connection(self, connection, id, host, port, parent_port)
+    def create_connection(self, connection: socket.socket, host: str, port: int, node_id=None, parent_port: int = None) -> Connection:
+        return Connection(self, connection, host, port, node_id, parent_port)
 
     def send_to_nodes(self, data: bytes, exclude: List[Connection] = None,
                       compression: bool = False) -> None:
@@ -69,7 +67,7 @@ class Node(threading.Thread):
         else:
             self.debug_print("send_to_node: node not found!")
 
-    def connect_with_node(self, host: str, port: int, reconnect: bool = False) -> bool:
+    def connect_with_node(self, host: str, port: int, reconnect: bool = False, our_node_id="") -> bool:
         if host == self.host and port == self.port:
             self.debug_print(
                 "connect_with_node: cannot connect with yourself!")
@@ -78,7 +76,7 @@ class Node(threading.Thread):
         for node in self.connections:
             if node.host == host and node.port == port or node.parent_port == port:
                 self.debug_print(
-                    f"connect_with_node: already connected with node: {node.id}")
+                    f"connect_with_node: already connected with node: {node.node_id}")
                 return True
 
         try:
@@ -99,20 +97,8 @@ class Node(threading.Thread):
             verification = sock.recv(4096)
             latency = time.time() - start_time
             proof = decrypt(verification)
-            proof, new_port = proof.split(b",")
-            sock.send(proof + b"," + f"{self.port}".encode())
-
-            # # Check if id is a valid rsa public key
-            # if authenticate_public_key(node_id) is False:
-            #     sock.send("closing connection: suspicious uid".encode())
-            #     self.debug_print(f"closing connection: suspicious uid: {node_id}")
-            #     sock.close()
-            #     return False
-
-            # Add a confirmation mechanism where we send randomized number encrypted via
-            # public key to validate identity
-            # rand_n = random.random()
-            # sock.send(())
+            proof, new_port, node_id = proof.split(b",")
+            sock.send(proof + b"," + f"{self.port}".encode() + b"," + our_node_id.encode())
 
             # # Close process if already connected / self (commented out to enable local testing)
             # if self.public_key == connected_node_id or connected_node_id in node_ids:
@@ -121,12 +107,12 @@ class Node(threading.Thread):
             #     return True
 
             # Form connection
-            thread_client = self.create_connection(sock, "test-ID", host, int(new_port), port)
+            thread_client = self.create_connection(sock, host, int(new_port), node_id, port)
             thread_client.start()
             thread_client.latency = latency
 
-            self.connections.append(thread_client)
-            # self.outbound.append(thread_client)
+            # self.connections.append(thread_client)
+            self.outbound.append(thread_client)
 
             # If reconnection to this host is required, add to the list
             if reconnect:
