+++ tests/test_dht.py
@@ -1,4 +1,4 @@
-from src.p2p.dht_node import DHTNode
+from src.p2p.smart_node import SmartDHTNode
 import time
 
 
@@ -7,25 +7,16 @@ if __name__ == "__main__":
     port = 5026
 
     # Spawn 3 workers on their own ports + threads
-    worker1 = DHTNode(host=ip, port=port, debug=True, wallet_address="5HDxH5ntpmr7U3RjEz5g84Rikr93kmtqUWKQum3p3Kdot4Qh",)
-    worker2 = DHTNode(host=ip, port=port + 1, debug=True, wallet_address="5HDxH5ntpmr7U3RjEz5g84Rikr93kmtqUWKQum3p3Kdot4Qh",)
-    worker3 = DHTNode(host=ip, port=port + 2, debug=True, wallet_address="5HDxH5ntpmr7U3RjEz5g84Rikr93kmtqUWKQum3p3Kdot4Qh",)
-    # worker4 = DHTNode(host=ip, port=port + 3, debug=True, wallet_address="5HDxH5ntpmr7U3RjEz5g84Rikr93kmtqUWKQum3p3Kdot4Qh",)
-    # worker5 = DHTNode(host=ip, port=port + 4, debug=True, wallet_address="5HDxH5ntpmr7U3RjEz5g84Rikr93kmtqUWKQum3p3Kdot4Qh",)
-    # worker6 = DHTNode(host=ip, port=port + 5, debug=True, wallet_address="5HDxH5ntpmr7U3RjEz5g84Rikr93kmtqUWKQum3p3Kdot4Qh",)
+    worker1 = SmartDHTNode(host=ip, port=port, debug=True, public_key="5HDxH5ntpmr7U3RjEz5g84Rikr93kmtqUWKQum3p3Kdot4Qh")
+    worker2 = SmartDHTNode(host=ip, port=port + 1, debug=True, public_key="5HDxH5ntpmr7U3RjEz5g84Rikr93kmtqUWKQum3p3Kdot4Qh")
+    worker3 = SmartDHTNode(host=ip, port=port + 2, debug=False, public_key="5HDxH5ntpmr7U3RjEz5g84Rikr93kmtqUWKQum3p3Kdot4Qh")
 
     worker1.start()
     worker2.start()
     worker3.start()
-    # worker4.start()
-    # worker5.start()
-    # worker6.start()
 
-    worker2.connect_with_node(ip, port)
-    worker3.connect_with_node(ip, port)
-    # worker4.connect_with_node(ip, port)
-    # worker5.connect_with_node(ip, port)
-    # worker6.connect_with_node(ip, port)
+    worker2.connect_dht_node(ip, port)
+    worker3.connect_dht_node(ip, port)
 
     time.sleep(15)
