+++ tests/test_workers.py
@@ -30,15 +30,15 @@ if __name__ == "__main__":
     # worker3.start()
 
     # Hard code workers connecting to the master node, ideally this will be done via smart contract or DHT
-    worker1.connect_with_node(ip, port + 1)
-    # worker1.connect_with_node(ip, port + 2)
+    worker1.connect_dht_node(ip, port + 1)
+    # worker1.connect_dht_node(ip, port + 2)
 
     dummy_input = torch.zeros((1, 1), dtype=torch.long)
     model = BertModel.from_pretrained("bert-base-uncased")
     # model = Wav2Vec2BertModel.from_pretrained("facebook/w2v-bert-2.0")
 
     time.sleep(10)
-    d_model = DistributedModel(model, worker1, torch.optim.Adam)
+    d_model = DistributedModel(model, worker1)
 
     # with open("distributed_graph.json", "w") as f:
     #     json.dump(graph, f, indent=4)
